Refactor package.scala to remove unused code and update default values
